+++ src/ast.c
@@ -96,6 +96,7 @@ Expr *expr_llamada(Arena *a, Expr *callee, Expr **args, int n_args, int linea, i
         e->como.llamada.args_spread = NULL;  /* v1.22: NULL = sin spreads */
         e->como.llamada.kwarg_keys = NULL;   /* v1.23: NULL = sin kwargs */
         e->como.llamada.kwarg_lens = NULL;
+        e->como.llamada.args_doble_spread = NULL;  /* v1.25: NULL = sin **spread */
     }
     return e;
 }